added seeding to challenges on production profile
